+++ NEWS.md
@@ -1,5 +1,14 @@
 Phan NEWS
 
+??? ?? 2019, Phan 2.2.1 (dev)
+-----------------------
+
+New features(Analysis):
++ Reduce the number of false positives of `--redundant-condition-detection` for variables in loops
+
+Bug fixes:
++ Fix issue that would make Phan infer that a redundant/impossible condition outside a loop was in a loop.
+
 Jun 14 2019, Phan 2.2.0
 -----------------------